Improve code modularity and add privilege summary
This is the output from our pairing session earlier today. It's still
a work in progress, as the data file has multiple missing and
placeholder elements, and we don't apply any useful formatting to the
final list of privileges. But it's a start at organizing the data and
letting us work with it in a more structured way.